grid/backtester_and_optimizer/requirements.txt
aiodns==3.0.0
aiofiles==23.2.1
aiohttp==3.8.5
aiohttp-socks==0.8.4
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
ccxt==4.2.53
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dateparser==1.1.8
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.12.4
frozenlist==1.4.0
greenlet==3.0.1
identify==2.5.30
lxml==4.9.3
multidict==6.0.4
mypy==1.5.1
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pandas-stubs==2.1.1.230928
pathspec==0.11.2
platformdirs==3.11.0
pre-commit==3.4.0
psycopg2==2.9.9
pybit==5.6.2
pycares==4.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
python-socks==2.4.4
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests
requests-futures==1.0.1
ruff==0.0.292
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
types-aiofiles==23.2.0.0
types-pytz==2023.3.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
urllib3==2.0.6
virtualenv==20.24.5
websocket-client==1.7.0
websockets==12.0
yarl==1.9.2
